solve

+++ src/main/java/com/cs4/CS4/model/Booking.java
@@ -7,11 +7,12 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.ManyToOne;
+import com.cs4.CS4.model.BookingStatus;
 @Entity
 public class Booking {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     @ManyToOne
     private User user;
     @ManyToOne
@@ -20,10 +21,10 @@ public class Booking {
     private LocalDate checkOut;
     @Enumerated(EnumType.STRING)
     private BookingStatus status;
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	public User getUser() {
